==> 1_about libraries.py <==
# we are using different libraries in data analysis

# Scientific computing libraries
# 1. pandas for data structure and tools
# 2. numpy for arrays and matrices
# 3. scipy for integrals solving differential equ. and optimization#

# visulizing libraries
# 1. matplotlib for plots & graphs
# 2. seaborn for heatmaps and timeseries
#
# algorithmic library used for ML
# 1. scikit-learn for ML: regression , classification, and clustering
# 2. statsmodels estimate statistical modules and perform statistical test
