==> 5_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

path="D:/my own/automobile.csv"

df = pd.read_csv(path )


df.columns

**Data Preprocessing**
process of converting or mapping data from one “raw” form into another also known as data cleaning or wrangling

you can add a value to each entry of a column.
To add 1 to each “symbolling” entry, use df["symboling"]=df["symboling"]+1.


df["symboling"]=df["symboling"]+1

**Dealing With Missing Values in Python**

When no data value is stored for a feature for a particular observation, we say this
feature has a “missing value”.
Usually, missing value in dataset appears as “?”, “N/A”, 0, "NaN" or just a blank cell.

*methods to deal the missing value*
- drop the variable or whole data entry
- replace it with an average(mean) of column
- replace it with the frequency (mode) of that column


**Remove the missing data**
To remove data that contains missing values, pandas library has a built-in method called
**‘dropna’**.
Essentially, with the dropna method, you can choose to drop rows or columns that contain
missing values, like NaN. So you’ll need to specify “axis=0” to drop the rows, or “axis=1” to drop the columns that contain the missing values.

It can simply be done in one line of code using dataframe.dropna().
Setting the argument “inplace” to “true” allows the modification to be done on the
dataset directly.
“Inplace=True” just writes the result back into the dataframe.

df.head(5)

now we have to delete the NaN values in normalized losses 

df.dropna(subset=["normalized-losses"], axis=0)

#df.dropna(subset=["normalized-losses"], axis=0, inplace = True)

this code will update the data frame

**Replacing the missing values**

To replace missing values like NaNs with actual values, pandas library has a built in method
called ‘replace’, which can be used to fill in the missing values with the newly calculated values.
"dataframe.replace(missing value, new value)"

assume that we want to replace the missing values of the variable ‘normalized-losses’ by the mean value of the variable.

mean=df["normalized-losses"].mean()
print(mean)

df.head(10)

df["normalized-losses"].replace(np.nan , mean)
