==> 6_data_Formating.ipynb <==
import pandas as pd
import numpy as np

path="D:/my own/automobile.csv"


df = pd.read_csv(path )


the Referring to our used car dataset, there’s a feature named “city-mpg” in the dataset,
which refers to a car fuel consumption in miles per gallon unit.
its column no. is 23 **".iloc[starting index : ending index , column no]"** this function is used to print a specific column data

df.iloc[:10, 23]

now there may be someone who lives in a country that uses metric units.
So you would want to convert those values to L/100km --the metric version.
To transform mpg to L/100km we need to divide 235 by each value in the city-mpg column.

df["city-mpg"]=235/df["city-mpg"]

df.rename(columns={"city-mpg" : "city-L/100km"}) #inplace=True )

To identify a features data type, in Python we can use the dataframe.dtypes() method and
check the datatype of each variable in a dataframe.
In the case of wrong datatypes, the method dataframe.astype() can be used to convert
a datatype from one format to another.

df["price"] = df["price"].astype(int)

df.dtypes["price"]
